fixed layout to ensure the buttons and pages can work for each
