+++ app/src/main/java/fragments/MainActivity.kt
@@ -9,9 +9,7 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import edu.uc.zhukv.droneradarmap.MapsActivity
 import edu.uc.zhukv.droneradarmap.R
-import edu.uc.zhukv.droneradarmap.Weather_Layer.TransparentTileOWM
 import kotlinx.android.synthetic.main.fragment_home.*
 import main.CurrentWeatherConditions
 
@@ -19,12 +17,11 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.fragment_home, R.layout.activity_main, R.layout.activity_maps, R.layout.details_fragment, R.layout.fragment_notifications,
-                R.layout.main_fragment)
+        setContentView(R.layout.activity_main)
 
         //enables back button for maps page
         val actionBar = supportActionBar
-        actionBar!!.title = "Maps Activity"
+        actionBar!!.title = "Go-4-Drone"
         actionBar.setDisplayHomeAsUpEnabled(true)
 
         //enables navigation bar
@@ -36,30 +33,6 @@ class MainActivity : AppCompatActivity() {
 
         bottomNavigationView.setupWithNavController(navController)
 
-        //enables current weather flag system details
-        imgCurrentFlag.setOnClickListener {
-            val intent = Intent(this, CurrentWeatherConditions::class.java)
-            startActivity(intent)
-        }
-
-        //enables see radar button
-        btnSeeRadar.setOnClickListener {
-            val intent = Intent(this, TransparentTileOWM::class.java)
-            startActivity(intent)
-        }
-
-        //enables change location button
-        btnChangeLocation.setOnClickListener {
-            val intent = Intent(this, MapsActivity::class.java)
-            startActivity(intent)
-        }
-
-        //enable the drop list for the map option
-        val adapter = ArrayAdapter.createFromResource(this, R.array.map_options, android.R.layout.simple_spinner_item)
-        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-    }
-
-    private fun setContentView(fragmentHome: Int, activityMain: Int, activityMaps: Int, detailsFragment: Int, fragmentNotifications: Int, mainFragment: Int) {
 
     }
